debug. 중요. - 자바에서는 /연산을 int끼리 사용하면 변수 타입이 double이더라도 int로 계산된 우측값을 갖는다. - 고로 명시적으로 double을 사용하고 싶다면 2.0처럼 직접적으로 소수점을 표시해야한다. - 문제에서 모든 꼭짓점이 정수라고 명시되어있지만, 계산과정에서 임의로 사용한 중점 역시 당연하게 정수로 판단하여 초기에 누락이 발생했다. - 로직에서 문제가 없다면 문제의 조건을 꼭 따져보고, 놓친 것 뿐 아니라 명시되지 않았는데 당연하게 적용한 것을 찾아내어야 한다.

+++ 백준 단계별로 풀어보기/10. 기하 1 - 직사각형과 삼각형/3009.java
@@ -31,17 +31,17 @@ class Main {
                     + get_distance_multiple(xs[i], ys[i], point3_x, point3_y);
             // 만약 현재 i번째가 직각점(피타고라스 법칙에 해당)이라면, 다른 두 점의 중점을 기준으로 직각점을 대칭.
             if(sum_right_sides == get_distance_multiple(point2_x, point2_y, point3_x, point3_y)){
-                int mid_x = (point2_x+point3_x)/2;
-                int mid_y = (point2_y+point3_y)/2;
+                double mid_x = (point2_x+point3_x)/2.0;
+                double mid_y = (point2_y+point3_y)/2.0;
                 int result_x, result_y;
                 if(mid_x<xs[i])
-                    result_x = mid_x - (xs[i]-mid_x);
+                    result_x = (int)(mid_x - (xs[i]-mid_x));
                 else
-                    result_x = mid_x + (mid_x-xs[i]);
+                    result_x = (int)(mid_x + (mid_x-xs[i]));
                 if(mid_y<ys[i])
-                    result_y = mid_y - (ys[i]-mid_y);
+                    result_y = (int)(mid_y - (ys[i]-mid_y));
                 else
-                    result_y = mid_y + (mid_y-ys[i]);
+                    result_y = (int)(mid_y + (mid_y-ys[i]));
                 StringBuilder sb = new StringBuilder();
                 sb.append(result_x);
                 sb.append(" ");
